move more stuff around for iport cycles
